requirements.txt
numpy>=1.19.0
matplotlib>=3.3.0
pillow>=8.0.0
transformers>=4.15.0
torch>=1.10.0 